Added STM32 <-> ESP32 communications on the dog
STM32 and ESP32 on the dog are connected via UART.
We want the ESP32 to send a 0xFF after it is ready.
We want the STM32 to send a random character as a request to get data
from the ESP32. When the ESP32 receives at least one character (any), it
will respond with a quaternion (4 floats == 16 bytes).